rename new-sankey to just sankey
